Fixed console_input.py keeping the process open
Using advice from stackoverflow (https://stackoverflow.com/questions/42288332/how-to-cleanly-exit-a-python-async-app-with-executor-tasks-still-waiting-for-con) I've removed the _python_exit() function from concurrent.futures.thread, and replaced it with a functionally identical version that doesn't try to join() daemon threads. This stops the process from hanging from the incomplete input() request, and because the thread is a daemon, it still terminates when the rest of the process terminates.